settings: convert each section to individual components
