+++ src/lib/components/FieldHelp.svelte
@@ -1,0 +1,9 @@
+<div class="field-help">
+	<slot />
+</div>
+
+<style lang="postcss">
+	.field-help {
+		@apply my-2 flex flex-col gap-y-3 px-0.5 text-muted;
+	}
+</style>
